+++ example/lib/main.local.dart
@@ -179,7 +179,7 @@ String testPath = '/deep/orgs/ACMEORG/dept/accounting/staff/mellisa/babesy//';
   router.log('GET $testPath Response: ${dres.statusCode} - ${dres.body}');
   router.log('\n');
 
-  // return;
+  return;
   // return;
 
   final removetRequest = mockRequest(path: '/context-example');